requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-groq
groq
streamlit
requests
pydantic
python-dotenv
joblib
